requirements.txt
numpy
scipy
pandas
fastparquet
celery[redis]
pulp

psycopg2-binary

django >= 4.0
djangorestframework
django-cors-headers
django-extensions
django-import-export
django-celery-beat
django-filter
djangorestframework-simplejwt
drf-spectacular
drf-spectacular-sidecar
Pillow
alibabacloud_dysmsapi20170525==2.0.23

gevent
gunicorn

krux
knob
dt_utils
python-dotenv
SQLAlchemy

# install manually
# ncore
# django_ncore
# datable
# work_order
# forecast
# optplan
# vali